Prepared for work with jar file and use

+++ src/main/java/ru/turlyunef/App.java
@@ -11,15 +11,28 @@ public class App {
 
     public static void main(String[] args) {
         log.info("Start the program");
-        Data data = new CableMagazineData();
-        ArrayList<Data> allData;
-        allData = FileConverter.convertFileToDataArrays("Resources/in.txt", data);
-        ArrayList<Data> result = Calculator.calculateCableMagazineData(allData, false);
-        File outputFile = new File(DataSaver.OUTPUT_FILE_NAME);
-        outputFile.delete();
-        DataSaver.saveSumResultToFile(result);
-        ArrayList<Data> resultForVOR = Calculator.calculateCableMagazineData(allData, true);
-        DataSaver.saveVORResultToFile(resultForVOR);
+        File myFolder = new File(".");
+        File[] files = myFolder.listFiles();
+        for (File x: files
+             ) {
+            if (x.getPath().contains(".csv")){
+                try {
+                    log.info("Calculating file: " + x.getPath());
+                    ArrayList<Data> allData;
+                    allData = FileConverter.convertFileToDataArrays(x.getPath(), new CableMagazineData());
+                    ArrayList<Data> result = Calculator.calculateCableMagazineData(allData, false);
+                    String[] names = x.getPath().split(".csv");
+                    File outputFile = new File(names[names.length-1] + "_result.txt");
+                    outputFile.delete();
+                    DataSaver.saveSumResultToFile(outputFile.getName(), result);
+                    ArrayList<Data> resultForVOR = Calculator.calculateCableMagazineData(allData, true);
+                    DataSaver.saveVORResultToFile(outputFile.getName(), resultForVOR);
+                }
+                catch (NullPointerException e){
+                    log.error("Error of calculating file: " + x.getPath());
+                }
+            }
+        }
         log.info("End the program");
     }
 }
